[refactor] store 들의 이름을 변경
