+++ src/common/components/TheHeader/TheHeader.vue
@@ -1,14 +1,14 @@
 <script setup lang="ts">
 import { storeToRefs } from 'pinia';
 
-import useUserStore from '@common/stores/user-store';
+import useUserModeStore from '@common/stores/user-mode-store';
 
-const userStore = useUserStore();
-const { userMode } = storeToRefs(userStore);
+const userModeStore = useUserModeStore();
+const { userMode } = storeToRefs(userModeStore);
 
 function changeUserMode() {
-  if (userMode.value === 'USER') userStore.updateUserMode('STAFF');
-  else userStore.updateUserMode('USER');
+  if (userMode.value === 'USER') userModeStore.updateUserMode('STAFF');
+  else userModeStore.updateUserMode('USER');
 }
 </script>
